+++ irs-api/src/main/java/org/eclipse/tractusx/irs/semanticshub/SemanticsHubClient.java
@@ -21,14 +21,25 @@
  ********************************************************************************/
 package org.eclipse.tractusx.irs.semanticshub;
 
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Base64;
 import java.util.Map;
+import java.util.Optional;
 
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.eclipse.tractusx.irs.configuration.RestTemplateConfig;
+import org.eclipse.tractusx.irs.services.validation.SchemaNotFoundException;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
 
@@ -43,7 +54,7 @@ interface SemanticsHubClient {
      * @param urn of the model
      * @return Json Schema
      */
-    String getModelJsonSchema(String urn);
+    String getModelJsonSchema(String urn) throws SchemaNotFoundException;
 
 }
 
@@ -66,6 +77,7 @@ class SemanticsHubClientLocalStub implements SemanticsHubClient {
  * Semantics Hub Rest Client Implementation
  */
 @Service
+@Slf4j
 @Profile({ "!local && !test" })
 class SemanticsHubClientImpl implements SemanticsHubClient {
 
@@ -73,15 +85,20 @@ class SemanticsHubClientImpl implements SemanticsHubClient {
 
     private final RestTemplate restTemplate;
     private final String semanticsHubUrl;
+    private final String localModelDirectory;
 
     /* package */ SemanticsHubClientImpl(
             @Qualifier(RestTemplateConfig.SEMHUB_REST_TEMPLATE) final RestTemplate restTemplate,
-            @Value("${semanticsHub.modelJsonSchemaEndpoint:}") final String semanticsHubUrl) {
+            @Value("${semanticsHub.modelJsonSchemaEndpoint:}") final String semanticsHubUrl,
+            @Value("${semanticsHub.localModelDirectory:}") final String localModelDirectory) {
         this.restTemplate = restTemplate;
         this.semanticsHubUrl = semanticsHubUrl;
+        this.localModelDirectory = localModelDirectory;
 
         if (StringUtils.isNotBlank(semanticsHubUrl)) {
             requirePlaceholder(semanticsHubUrl);
+        } else if (StringUtils.isBlank(localModelDirectory)) {
+            log.warn("No Semantic Hub URL or local model directory was provided. Cannot validate submodel payloads!");
         }
     }
 
@@ -98,9 +115,40 @@ class SemanticsHubClientImpl implements SemanticsHubClient {
     }
 
     @Override
-    public String getModelJsonSchema(final String urn) {
-        final UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromUriString(semanticsHubUrl);
-        final Map<String, String> values = Map.of(PLACEHOLDER_URN, urn);
-        return restTemplate.getForObject(uriBuilder.build(values), String.class);
+    public String getModelJsonSchema(final String urn) throws SchemaNotFoundException {
+        return readFromSemanticHub(urn).or(() -> readFromFilesystem(urn))
+                                       .orElseThrow(() -> new SchemaNotFoundException(
+                                               "Could not load model with URN " + urn));
+    }
+
+    private Optional<String> readFromFilesystem(final String urn) {
+        if (StringUtils.isNotBlank(localModelDirectory)) {
+            final Path path = Paths.get(localModelDirectory, normalize(urn));
+            if (path.toFile().exists()) {
+                try {
+                    return Optional.of(Files.readString(path));
+                } catch (IOException e) {
+                    log.error("Unable to read schema file at path '{}'", path, e);
+                }
+            }
+        }
+        return Optional.empty();
+    }
+
+    private Optional<String> readFromSemanticHub(final String urn) {
+        if (StringUtils.isNotBlank(semanticsHubUrl)) {
+            try {
+                final UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromUriString(semanticsHubUrl);
+                final Map<String, String> values = Map.of(PLACEHOLDER_URN, urn);
+                return Optional.ofNullable(restTemplate.getForObject(uriBuilder.build(values), String.class));
+            } catch (final RestClientException e) {
+                log.error("Unable to retrieve schema from semantic hub for urn '{}'", urn, e);
+            }
+        }
+        return Optional.empty();
+    }
+
+    private String normalize(final String urn) {
+        return Base64.getEncoder().encodeToString(FilenameUtils.getName(urn).getBytes(StandardCharsets.UTF_8));
     }
 }